pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "claw401"
version = "0.1.0"
description = "Python SDK for the Claw401 X401 wallet authentication protocol"
authors = ["Claw401 Protocol <protocol@claw401.xyz>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://claw401.xyz"
repository = "https://github.com/claw401"
keywords = ["solana", "authentication", "wallet", "web3", "x401", "agent", "identity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Internet :: WWW/HTTP :: Session",
]
packages = [{ include = "claw401" }]

[tool.poetry.dependencies]
python = "^3.10"
PyNaCl = "^1.5.0"
base58 = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.4"
mypy = "^1.13.0"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
strict = true
python_version = "3.10"
